+++ frontend/middleware/driver/web/demo.html
@@ -19,6 +19,7 @@
 
         .header {
             background-color: rgba(255, 255, 255, 0.1);
+            -webkit-backdrop-filter: blur(10px);
             backdrop-filter: blur(10px);
             color: white;
             padding: 1em 0;
@@ -148,8 +149,13 @@
         }
 
         @keyframes spin {
-            0% { transform: rotate(0deg); }
-            100% { transform: rotate(360deg); }
+            0% {
+                transform: rotate(0deg);
+            }
+
+            100% {
+                transform: rotate(360deg);
+            }
         }
 
         .status-success {
@@ -162,6 +168,7 @@
 
         .footer {
             background-color: rgba(255, 255, 255, 0.1);
+            -webkit-backdrop-filter: blur(10px);
             backdrop-filter: blur(10px);
             color: white;
             text-align: center;
@@ -230,7 +237,7 @@
         async function testProtectedAPI() {
             const resultElement = document.getElementById('apiResult');
             resultElement.innerHTML = '<div class="loading"></div>APIをテスト中...';
-            
+
             try {
                 const response = await fetch('/api/protected', {
                     method: 'GET',
@@ -238,7 +245,7 @@
                 });
 
                 const data = await response.json();
-                
+
                 if (response.ok) {
                     resultElement.innerHTML = JSON.stringify(data, null, 2);
                     resultElement.className = 'api-result status-success';
@@ -278,7 +285,7 @@
         function displayUserInfo(data) {
             const loginTime = new Date(data.loginAt).toLocaleString('ja-JP');
             const currentTime = new Date().toLocaleString('ja-JP');
-            
+
             content.innerHTML = `
                 <h2>✅ 認証成功</h2>
                 
@@ -342,7 +349,7 @@
                     <a href="/" class="btn btn-primary">🏠 ログインページへ</a>
                 </div>
             `;
-            
+
             // 3秒後に自動でリダイレクト
             setTimeout(() => {
                 window.location.href = '/';
@@ -365,7 +372,7 @@
                 const response = await fetch('/user/auth', {
                     credentials: 'include'
                 });
-                
+
                 if (!response.ok) {
                     displayAuthError();
                 }
